% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{desUniformInit}
\alias{desUniformInit}
\title{Uniformly distributed initial population}
\usage{
desUniformInit(fitnessFcn, nvars, control)
}
\arguments{
\item{fitnessFcn}{Des algorithm fitness function. Not used in this
init, but might be useful in a custom function}

\item{nvars}{Amount of variables in the optimized problem}

\item{control}{A control list from the des algorithm. Important
elements are control$upperBounds and control$lowerBounds,
specyfying the initial search space, and control$popSize defining
the amount of generated individuals.}
}
\value{
A numeric matrix with dimensions (control$popSize x nvars),
with each row containing subsequent variables of each individual
}
\description{
Returns initial populations for the des algorithm using
uniform distribution.
}
